drops: optimize item cleanup process.
the idea here is to split the "drops" system into two different processes. we have an async task that is "querying" item information, and updating queues so that a sync task can perform the actual removal. doing all of this at once on a sync task can harm TPS pretty bad once 10k+ items exist in the world.